requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.4
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.5
waitress==2.1.2
whitenoise==6.2.0
